requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# OpenAI
openai>=1.12.0

# PDF Processing
pypdf>=4.0.1

# Vector Database
pinecone>=5.0.0

# LangChain for RAG (updated for Python 3.13 compatibility)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# Text Processing
tiktoken>=0.5.2

# Environment Variables
python-dotenv>=1.0.1

# Additional utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
